Experimenting with local servers. May need to revert.

+++ project/Assets/Plugins/Photon Unity Networking/Plugins/PhotonNetwork/RoomInfo.cs
@@ -10,6 +10,7 @@
 // ----------------------------------------------------------------------------
 
 using System;
+using UnityEngine;
 using ExitGames.Client.Photon;
 
 
@@ -195,7 +196,7 @@ public class RoomInfo
         // fetch the "well known" properties of the room, if available
         if (propertiesToCache.ContainsKey(GameProperties.MaxPlayers))
         {
-            this.maxPlayersField = (byte)propertiesToCache[GameProperties.MaxPlayers];
+			this.maxPlayersField = (byte)(int)propertiesToCache[GameProperties.MaxPlayers];
         }
 
         if (propertiesToCache.ContainsKey(GameProperties.IsOpen))
